requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
psycopg2-binary
pyyaml
requests
django-cors-headers
djangorestframework_simplejwt
python-decouple